% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_draft.R
\name{assemble_draft}
\alias{assemble_draft}
\title{Deprecated function for draft import}
\usage{
assemble_draft(...)
}
\arguments{
\item{...}{A named vector of filenames which will be imported as
patterns stored in the returned draft, with the names used as indices.
Files should be plain text.}
}
\value{
Returns a list (the same length as ...)
containing the imported patterns.
}
\description{
assemble_draft has been deprecated (as of development version 0.4.2) in
favor of import_draft, which has the same semantics (and is actually now
the same code -- assemble_draft is now only an alias for import_draft.) This
should hopefully make the link between import_draft and import_pattern clear,
and more importantly distinguish these functions from extract_pattern and
the new extract_draft function.
}
\examples{
\dontrun{
export_template("my sample pattern", "out.Rmd")
assemble_draft("sample_pattern" = "out.Rmd")
}

}
