% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_yaml_header.R
\name{create_yaml_header}
\alias{create_yaml_header}
\title{Convert list objects into R Markdown YAML headers}
\usage{
create_yaml_header(..., line.sep = c("\\n", "\\r\\n", "\\r"),
  indent = 2, omap = FALSE, column.major = TRUE, unicode = TRUE,
  precision = getOption("digits"), indent.mapping.sequence = FALSE,
  handlers = NULL)
}
\arguments{
\item{...}{A set of objects that will be combined into the YAML header.
Objects may be provided as lists (the structure
list("outputs" = "html_document") translates to outputs: html_document)
or as single-item named vectors (passing "title" = "My Report" to ... will
translate to title: "My Report").}

\item{line.sep, indent, omap, column.major, unicode, precision, indent.mapping.sequence, handlers}{Additional arguments to be passed to [yaml::as.yaml]}
}
\value{
A string, formatted for use as an R Markdown YAML header
}
\description{
This function tweaks the behavior of [yaml::as.yaml] to return a string
which can immediately be used as an R Markdown YAML header. It's designed
to accept both deeply nested lists and simpler list formats to make
reasoning about your header easier.
}
