% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heddlr-package.R
\docType{package}
\name{heddlr-package}
\alias{heddlr}
\alias{heddlr-package}
\title{heddlr: Dynamic R Markdown Document Generation}
\description{
Helper functions designed to make
    dynamically generating R Markdown documents easier by providing a
    simple and tidy way to create report pieces, shape them to your data,
    and combine them for exporting into a single R Markdown document.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mikemahoney218/heddlr}
  \item \url{https://mikemahoney218.github.io/heddlr/}
  \item Report bugs at \url{https://github.com/mikemahoney218/heddlr/issues}
}

}
\author{
\strong{Maintainer}: Michael Mahoney \email{mike.mahoney.218@gmail.com} (\href{https://orcid.org/0000-0003-2402-304X}{ORCID})

}
\keyword{internal}
