% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_draft.R
\name{assemble_draft}
\alias{assemble_draft}
\title{Assemble multiple patterns into a single draft object}
\usage{
assemble_draft(...)
}
\arguments{
\item{...}{A named vector of filenames which will be imported as
patterns stored in the returned draft, with the names used as indices.
Files should be plaintext.}
}
\value{
A list ("draft") object containing the imported patterns.
}
\description{
When working with multiple patterns that will be woven into a tetmplate,
it makes sense to have all patterns stored in a central object. This
function creates that object from a named vector of filenames to be used
in further generation, importing the files via [import_pattern].
}
