% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_template.R
\name{export_template}
\alias{export_template}
\title{Safely export templates to file.}
\usage{
export_template(template, filename, sep = "", filename.is.string = T)
}
\arguments{
\item{template}{The template string to be written out}

\item{filename}{The path to write the template to, passed to [base::writeLines].
Also accepts `stdout()` (and likely other similar functions) with a warning}
}
\description{
This is a simple wrapper function around [utf8::utf8_encode] and
[base::writeLines], letting users write their template strings to file without
having to worry about file encodings. For more details on why UTF-8
encoding is necessary, check out
[Yihui Xie's](https://yihui.org/en/2018/11/biggest-regret-knitr/) post on
the subject.
}
\details{
Note that this function is effectively the inverse of [import_pattern] --
export_template(import_pattern("out.txt"), "out.txt") should always result
in an unchanged file, and exceptions to this rule would be considered bugs.
}
